requirements.txt
PyMySQL ~= 1.0.2
cryptography ~= 38.0.1
mysql-connector-python ~= 8.0.31